fix: added multiple curric partners to the docx back page
Also refactored to pull out portableTextToDocx(...)
